Cleanup webpack (#48)
* Cleanup webpack config

* Move around a bunch of webpack files

* rename certain webpack files
