+++ lib/utils/compact-array.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function compactArray(array) {
+  return array.filter(Boolean);
+};
