+++ lib/webpack-compiler/create-webpack-compiler.js
@@ -3,11 +3,8 @@
 const webpack = require('webpack');
 
 const normalizeWebpackError = require('./normalize-webpack-errors');
-const createWebpackConfig = require('./create-webpack-config');
 
-module.exports = function createWebpackCompiler(urc) {
-  const webpackConfig = createWebpackConfig(urc);
-
+module.exports = function createWebpackCompiler(webpackConfig) {
   try {
     return webpack(webpackConfig);
   } catch (error) {
