+++ lib/webpack-compiler/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const writeWebpackStats = require('./write-webpack-stats');
-const webpackPromise = require('./webpack-promise');
+const webpackRunPromise = require('./webpack-run-promise');
 const normalizeWebpackError = require('./normalize-webpack-errors');
 const WebpackCompileError = require('./webpack-compile-error');
 const createWebpackCompiler = require('./create-webpack-compiler');
@@ -10,6 +10,6 @@ module.exports = {
   createWebpackCompiler,
   WebpackCompileError,
   writeWebpackStats,
-  webpackPromise,
+  webpackRunPromise,
   normalizeWebpackError
 };
