+++ lib/webpack-compiler/webpack-run-promise.js
@@ -1,14 +1,11 @@
 'use strict';
 
-const webpack = require('webpack');
 const normalizeWebpackError = require('./normalize-webpack-errors');
-const createWebpackConfig = require('./create-webpack-config');
+const createWebpackCompiler = require('./create-webpack-compiler');
 
-function webpackPromise(urc) {
-  const webpackConfig = createWebpackConfig(urc);
-
+function webpackRunPromise(webpackConfig) {
   return new Promise((resolve, reject) => {
-    const compiler = webpack(webpackConfig);
+    const compiler = createWebpackCompiler(webpackConfig);
     compiler.run((error, stats) => {
       const normalizedError = normalizeWebpackError({
         error,
@@ -19,10 +16,9 @@ function webpackPromise(urc) {
         reject(normalizedError);
         return;
       }
-
       resolve(stats);
     });
   });
 }
 
-module.exports = webpackPromise;
+module.exports = webpackRunPromise;
