+++ lib/webpack-config/safe-webpack-uglifier.js
