+++ lib/webpack-plugins/module-not-found-plugin.js
